consolidate browse judoka navigation tests
